Custom UUIDGenerator removed; Moved DatabaseExecutionContext to Common package
