+++ rest-api/app/common/DatabaseExecutionContext.java
@@ -1,4 +1,4 @@
-package models.utils;
+package common;
 
 import akka.actor.ActorSystem;
 import play.libs.concurrent.CustomExecutionContext;
