--- rest-api/app/models/utils/UUIDGenerator.java
@@ -1,6 +1,0 @@
-package models.utils;
-
-import org.hibernate.annotations.GenericGenerator;
-
-@GenericGenerator(name = "UUID", strategy = "uuid2")
-public @interface UUIDGenerator {}
