+++ rest-api/app/repositories/JPARepository.java
@@ -2,7 +2,7 @@ package repositories;
 
 import common.Constants;
 import lombok.AllArgsConstructor;
-import models.utils.DatabaseExecutionContext;
+import common.DatabaseExecutionContext;
 import play.db.jpa.JPAApi;
 
 import javax.persistence.EntityManager;
